[Add] Try auto add bot_already_beaten (client,tools)

+++ client/base/Api_protocol.h
@@ -415,7 +415,7 @@ public:
     void tryEscape();
     void useSkill(const uint16_t &skill);
     void heal();
-    void requestFight(const uint32_t &fightId);
+    void requestFight(const uint16_t &fightId);
     void changeOfMonsterInFightByPosition(const uint8_t &monsterPosition);
 
     //monster
